Registrar llamadas no precargadas con registrar citas y registrar en tabla de llamadas programadas

+++ Controllers/Usuarios/EliminarUsuario.js
@@ -1,45 +1,64 @@
 Eliminarusuario = (Id_usuario) => {
 
-    $.ajax({
-        url: `${URL}/Usuarios/${Id_usuario}`,
-        type: 'delete',
-        datatype: 'json',
-        success: function (datos) {
-
-            if(datos.data.Eliminar){
-
-                RecargarDataTable();
-
-            }else{
-
-                 swal({
-                    title: "Error al eliminar.",
-                    type: "error",
-                    text: "No se puede eliminar el usuario, debido a que está asociado a un registro de llamada o visita externa.",
-                    showCancelButton: false,
-                    confirmButtonColor: "#2F6885",
-                    confirmButtonText: "Cerrar",
-                    closeOnConfirm: true,
-                });
-
-            }
-           
-
-        },
-        error: function (error) {
-
-            swal({
-                title: "Error al eliminar.",
-                type: "error",
-                text: "No se pudo realizar la petición, intente más tarde.",
-                showCancelButton: false,
-                confirmButtonColor: "#2F6885",
-                confirmButtonText: "Cerrar",
-                closeOnConfirm: true,
+    wal({
+        title: "¿Desea eliminar este registro?",
+        text: "El registro será eliminado y no se podrá recuperar!",
+        type: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d9534f",
+        confirmButtonText: "Eliminar",
+        cancelButtonText: "Cancelar",
+        closeOnConfirm: false,
+        closeOnCancel: true
+      },
+      function(isConfirm) {
+        if(isConfirm){
+            $.ajax({
+                url: `${URL}/Usuarios/${Id_usuario}`,
+                type: 'delete',
+                datatype: 'json',
+                success: function (datos) {
+        
+                    if(datos.data.Eliminar){
+        
+                        RecargarDataTable();
+        
+                    }else{
+        
+                         swal({
+                            title: "Error al eliminar.",
+                            type: "error",
+                            text: "No se puede eliminar el usuario, debido a que está asociado a un registro de llamada o visita externa.",
+                            showCancelButton: false,
+                            confirmButtonColor: "#2F6885",
+                            confirmButtonText: "Cerrar",
+                            closeOnConfirm: true,
+                        });
+        
+                    }
+                   
+        
+                },
+                error: function (error) {
+        
+                    swal({
+                        title: "Error al eliminar.",
+                        type: "error",
+                        text: "No se pudo realizar la petición, intente más tarde.",
+                        showCancelButton: false,
+                        confirmButtonColor: "#2F6885",
+                        confirmButtonText: "Cerrar",
+                        closeOnConfirm: true,
+                    });
+                    console.log(error);
+                }
             });
-            console.log(error);
         }
-    });
+      })
+
+
+
+   
 
 }
